+++ src/Type/Collection/Validator/MaxAmountValidator.php
@@ -14,20 +14,14 @@ class MaxAmountValidator implements ValidatorInterface
      */
     private $config;
 
-    public function __construct(int $maxAmount, bool $negated=false, bool $dumpable=true)
+    public function __construct(int $maxAmount, bool $dumpable=true)
     {
-        $this->config = new MaxAmountValidatorConfig($maxAmount, $negated, $dumpable);
+        $this->config = new MaxAmountValidatorConfig($maxAmount, $dumpable);
     }
 
-    /**
-     * @param mixed $value
-     * @param null $key
-     * @param CollectionInterface|null $collection
-     * @throws Exception\AmountValidatorException
-     */
     public function validate($value, $key = null, CollectionInterface $collection = null): void
     {
-        $this->config->isNegated() ? $this->assertFalse($value, $key, $collection) : $this->assertTrue($value, $key, $collection);
+        $this->assertTrue($value, $key, $collection);
     }
 
     public function assertTrue($value, $key = null, CollectionInterface $collection = null): void
@@ -37,16 +31,6 @@ class MaxAmountValidator implements ValidatorInterface
         }
 
         $msg = "Items in this collection can NOT be more than: {$this->config->getMaxAmount()}";
-        throw new Exception\AmountValidatorException($msg);
-    }
-
-    public function assertFalse($value, $key = null, CollectionInterface $collection = null): void
-    {
-        if((count($collection) + 1) > $this->config->getMaxAmount()){
-            return;
-        }
-
-        $msg = "Items in this collection can be more than: {$this->config->getMaxAmount()}";
         throw new Exception\AmountValidatorException($msg);
     }
 
@@ -69,7 +53,7 @@ class MaxAmountValidator implements ValidatorInterface
         /**
          * @var MaxAmountValidatorConfig $config
          */
-        return new self($config->getMaxAmount(), $config->isNegated(), $config->isDumpable());
+        return new self($config->getMaxAmount(), $config->isDumpable());
     }
 
     /**