+++ src/Type/Collection/Validator/MinimumAmountValidator.php
@@ -14,20 +14,14 @@ class MinimumAmountValidator implements ValidatorInterface
      */
     private $config;
 
-    public function __construct(int $minAmount, bool $negated=false, bool $dumpable=true)
+    public function __construct(int $minAmount, bool $dumpable=true)
     {
-        $this->config = new MinimumAmountValidatorConfig($minAmount, $negated, $dumpable);
+        $this->config = new MinimumAmountValidatorConfig($minAmount, $dumpable);
     }
 
-    /**
-     * @param mixed $value
-     * @param null $key
-     * @param CollectionInterface|null $collection
-     * @throws Exception\AmountValidatorException
-     */
     public function validate($value, $key = null, CollectionInterface $collection = null): void
     {
-        $this->config->isNegated() ? $this->assertFalse($value, $key, $collection) : $this->assertTrue($value, $key, $collection);
+        $this->assertTrue($value, $key, $collection);
     }
 
     public function assertTrue($value, $key = null, CollectionInterface $collection = null): void
@@ -37,17 +31,6 @@ class MinimumAmountValidator implements ValidatorInterface
         }
 
         $msg = "Items in this collection must be at least: {$this->config->getMinAmount()}";
-
-        throw new Exception\AmountValidatorException($msg);
-    }
-
-    public function assertFalse($value, $key = null, CollectionInterface $collection = null): void
-    {
-        if((count($collection) - 1) <= $this->config->getMinAmount()){
-            return;
-        }
-
-        $msg = "Items in this collection must NOT be at least: {$this->config->getMinAmount()}";
 
         throw new Exception\AmountValidatorException($msg);
     }
@@ -71,7 +54,7 @@ class MinimumAmountValidator implements ValidatorInterface
         /**
          * @var MinimumAmountValidatorConfig $config
          */
-        return new self($config->getMinAmount(), $config->isNegated(), $config->isDumpable());
+        return new self($config->getMinAmount(), $config->isDumpable());
     }
 
     /**